Moved fsa layers to components/layers
